# api/src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  title: Сервис по генеалогии - OpenAPI спецификация
  description: Сервис позволяет зарегистрированному пользователю создавать генеалогическое древо
  version: 0.0.1
servers:
  - url:  https://localhost:8686/genealogy/swagger-ui/index.html
    description: Dev Server
paths:
...
  /person/{id}/parents/:
    post:
      summary: Метод добавления человека в генеалогическое древо
      description: Метод предназначен для добавления в БД человека
      tags:
        - Parents
      parameters:
        - name: id
          required: true
          in: path
          description: Идентификатор карточки человека
          example: 42abcd2b-8b9c-4af9-88f7-0bc180cf74b4
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: '#/components/schemas/Parents'
      responses:
        '201':
          description: Подтверждение успешного сохранения
          content:
            application/json;charset=UTF-8:
              schema:
                $ref: '#/components/schemas/Parents'
        '404':
          description: Отсутсвует сущность с данным id
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Error'
        '400':
          description: Некорректные входные данные. Возвращает список атрибутов с ошибками
          content:
            application/json;charset=UTF-8:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Error'
        '5XX':
          description: Любая неожиданная ошибка сервера
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Error'
  /person/{id}/parents/{parent_id}/:
    delete:
      summary: Метод удаления связи с человеком
      description: Метод предназначен для удаления связи с предком
      tags:
        - Parents
      parameters:
        - name: id
          required: true
          in: path
          description: Идентификатор карточки родственника
          example: 42abcd2b-8b9c-4af9-88f7-0bc180cf74b4
          schema:
            type: string
            format: uuid
        - name: parent_id
          required: true
          in: path
          description: Идентификатор карточки предка
          example: 42abcd2b-8b9c-4af9-88f7-0bc180cf74b4
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Успешное удаление
          content: {}
        '404':
          description: Отсутсвует сущность с данным id
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Error'
        '5XX':
          description: Любая неожиданная ошибка сервера
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    Parents:
      type: object
      required:
        - parentType
        - parent_id
      properties:
        id:
          type: string
          format: uuid
          example: 42abcd2b-8b9c-4af9-88f7-0bc180cf74b4
          description: Идентификатор карточки родителя
        parentType:
          $ref: '#/components/schemas/ParentType'
          example: mother
...
    ParentType:
      description: тип родителя
      type: string
      enum:
        - mother
        - father
      example: mother
    Error:
      type: object
      required:
        - title
        - detail
      properties:
        title:
          description: Краткое описание проблемы, понятное человеку
          type: string
          example: Entity not found
        detail:
          description: Описание конкретно возникшей ошибки, понятное человеку
          type: string
          example: Entity [Post] with id = [123456] not found. You MUST use POST to add entity instead of PUT
            
        errorCode:
          description: Код ошибки по бизнес логике приложения
          type: integer
          format: int32
          example: 57
            
        exception:
          description: Тип ошибки
          type: string
          example: NullPointerException
        stacktrace:
          description: Stacktrace в случае возникновения на сервере исключения
          type: string
          example: ...